requirements.txt
asgiref==3.4.1
Django==3.2.5
django-extensions==3.1.3
django-utils-six==2.0
gitdb==4.0.7
GitPython==3.1.18
Pillow==8.3.1
pydot==1.4.2
pyparsing==2.4.7
pytz==2021.1
smmap==4.0.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
user-agents==2.2.0
